improvements to webalizer script

+++ scripts/webalizer/processlogs.sh
@@ -30,13 +30,16 @@ do
 	# rename the files so that they are unique
 	for PREF in $PREFIXES
 	do
-		mv ${PREF}-${yyyymmdd} ${PREF}-${MACH}.log
+		if [ -e ${PREF}-${yyyymmdd} ]; then
+			mv ${PREF}-${yyyymmdd} ${PREF}-${MACH}.log
+		fi
 	done
 done
 
 for PREF in $PREFIXES
 do
-	wc -l ${PREF}-${MACH}.log
+	echo "============== $PREF ============="
+	wc -l ${PREF}-*.log
 	csh -c "(/usr/local/bin/mergelog ${PREF}-*.log > combined-${PREF}.log) >& /dev/null"	
 	rm -f ${PREF}-*.log
 done
@@ -63,8 +66,10 @@ rm -f agclimate.log
 # Step 4, archive these files
 for PREF in $PREFIXES
 do
-	mv combined-${PREF}.log ${PREF}-${yyyymmdd}.log
-	gzip ${PREF}-${yyyymmdd}.log
+	if [ -e combined-${PREF}.log ]; then
+		mv combined-${PREF}.log ${PREF}-${yyyymmdd}.log
+		gzip ${PREF}-${yyyymmdd}.log
+	fi
 done
 
 lftp -u akrherz@iastate.edu ftps://ftp.box.com << EOM
